debug the file BCVTemplate:
There was a call to a function LALCreateCoarseBank filling list1 and nlist variable.
Then, there was a call to LALCreateCoarseBank filling list2 and nlist (again) variable.

The bug was that later on, in a loop the code was trying to print the results of list1
with an iteration equal to nlist (which corresponded to list2).

Moreover all openned files in the following files where not closed at the end !!!