Changes to the InspiralWaveOverlapOut structure to include extra REAL4Vectors for returning xcorr1, xcorr2, filter1 and filter2.
Changes to the InspiralWaveOverlapIn to include a UINT2 flag 'ifExtOutput'.
When this flag is set to zero, the extra overlap output structures are not
filled in the call to LALInspiralWaveOverlap. If it is set to 1, then the
extra REAL4Vector items in InspiralWaveOverlapOut (see above) are filled
out.

Extra data structures and function prototypes for new subroutines
LALDiscoverInspiralEvents
LALEstimateEffectiveDistance
LALEvaluateInspiralChisqTest
LALTruncateInvSpectrum