Fix PR48:  correctly handle non-ASCII documents.
This change converts the Tokenizer class' internal buffer from an array of
ASCII characters to an array of wchar_t characters.  Python unicode strings
are added as one of the allowed types as which a token can be extracted.
All LIGO Light Weight string types in types.py except for the ilwd:char
variants are converted from str to unicode.