Major additions to ComputePSD: * choices for averaging over SFTs: sum, mean, median, harmonic sum, harmonic median, min, max: default is mean for backward compat. * same choices for averaging over IFOs: same options, default is harmonic *sum* for backward compat. * now you can bin the output of the final PSD, if you don't want 1/Tsft resolution. Options for size of bins, step of bins   (so you could have overlapping bins), and how to average the 1/Tsft bins into the larger output bins: same options as above
A couple of tests (PSDs of SFTs from 1 IFO / 2 IFOs) have been done and show this version agrees with the previous version
(floating point numerics are obviously not identical, but the octave plots line up perfectly). If you use this code please
test it against results you've used before to see if they agree, and try out the new stuff!