fix parallel build
lal-version needs liblal.la before it can be linked. This dependency was
originally specified as a complete path to liblal.la, however make does
not know what directory it is currently running in and therefore can't
determine that this path was the current directory. Using a relative
path therefore enables make to correctly calculate the dependencies,
leading to a successfull parallel build.