fix parallel build of lalapps
Similar to the fix in c16ca3256ea085ec90698d4c59a8636f76de5b7a the path
to the liblalapps.la library was absolute, this was therefore confusing
make is it believed that the library was not in the current directory
which in turn caused to the link dependencies being incorrect. Using a
relative path for the library fixes the parallel build.
(cherry picked from commit 7553159d4fbe1ab1f902ecb25c94b4c698e3222e)