top-level lalsuite configure functionality
Adds functionality so that you can configure both lal and lalapps from a
top level configure script.  There is a 00boot that calls lal/00boot
and lalapps/00boot and then there is a top level configure which will
configure both lal and lalapps.  The top level Makefile then makes both
lal and lalapps so that they can be installed together.  The lalapps
configure script is significantly modified to allow this.