use (LAL_)AC_PROG_CC_C99 macro for detecting c99 compiler
Versions of autoconf prior to 2.60 do not include the AC_PROG_CC_C99
macro and therefore are not as thorough at detecting a c99 compilant
compiler as more current versions. This patch enables the AC_PROG_CC_C99
macro to be used to detect the compiler if it is available or the local
version of this macro, LAL_AC_PROG_CC_C99, if it is not. This therefore
enables a C99 compilant compiler to be correctly detected on systems
running older versions of autoconf, such as CentOS 5.
(cherry picked from commit 71362ddaf4677ee1fe81c3547ded64b3c791fd7d)