Fix PR2533:
LIGOTimeGPS class in lal module is not hashable

This patch makes the seconds and nanoseconds attributes of LIGOTimeGPS
instances read-only, which effectively makes LIGOTimeGPS objects immutable
and therefore suitable for use as dictionary keys.  The patch, therefore,
adds a __hash__() method to the class.

This patch also improves the accuracy of the __mul__() and __div__()
operators, although they are still not in agreement with LAL.