Set new=False for new numpy.histogram API.
You get lots of warnings, but the behavior is identical to what it was before the upgrade. In the long term, we should adjust to new=True; in the new scheme, the right-most bin edge provided serves as a hard right edge instead of the left edge of a catch-all outlier bin.
(cherry picked from commit 620d37c6e172a94ef29403b3168cd8825949586c)