fix glitch in compute_thinca_livetimes()
In the recent patch to return the live time separately for each offset
vector, the return value in the "an instrument that must be off is never
vetoed" special case was not updated, it was still being returned as a
single 0.0 float instead of a list of 0.0 floats.  This fixes that.