Significant changes for performance.
The original plan had been to run segments_from_cats to generate
category files, then use this program to do the vetoes.  That turned
out to not be quite fast enough to process a minute of triggers in
a minute, so this now reads the veto_definer file directly and does
the filtering in a single query.