Don't get more DQ files than we need.
If there has been a lengthy gap in triggers lots of DQ files will
have accumulated in the meantime.  There's no need to load them all
and trying to do so may take more than the one minute this has to
complete.  So only load DQ files that overlap the trigger files.