initial packaging of liblalstochastic
this is the initial attempt at splitting the stochastic code off into a
separate package that can be released independently. all the code and
tests have been moved over but the documentation has not been integrated
into the build as it needs to be transitioned over to use doxygen
instead of laldoc.