adding the ability of rejecting some specific tags from the glob of input files in followup pipeline
I add a new field in section followup-triggers of the followup_pipe.ini file where one can specify a list of strings that should be excluded from the glob of input files. This patch is specially relevant for the followups of the S5 LV candidates, for which we need to be able to exclude times of short durations.