Revert "another attempt to make this REAL4-safe: use refTime=startTime for each segment"
This reverts commit ad7bc563c6520f648440a7b3e6f3eca7981d7f33.

This seems to actually hurt Hough-valdidation for some reason, while the Hough-tests
seem to pass without it. So for now we try to go with the previous version.

Note: there is nothing *wrong* with this version, other than it answers differ more from
the previous version than the validator allows. However, one can consider using it when
starting a new run, and validation against previous concurrent results is not a constraint.