move DataSegment struct definition to FindChirpDatatypes.h
The definition of the the DataSegment structure has been moved from the
DataBuffer.h header in the framedata package into the FindChirp.h header
as the findchirp package is the only package in LAL that uses this
structure. In fact the findchirp package was the only package in LAL
that used the DataBuffer.h header file.