add code to allow the ringdown pipeline make multiple different types of inj runs
The ringdown code needs the type of the injection to be specified, so that the correct injecting funtion can be called; LALRingInjectSignals for ringdown injections and LALFindChirpInjectSignals for EOBNR and Phenom. This code reads in the injection type from the [injection] section of the ihope ini file and places it in the [inspiral] section of the the hipe ini run for each injection. In addition, the executable needs to be changed. Ths solution to this is to have lalapps_inspinj as the default, but if the injection type is ringdown, this is replaced with lalapps_rinj.