add a script to write the log file of wscan_background pipeline
we want that the wscan_background pipeline generates a log file containing the start time and the end time of the pipeline. This will be done by this python script that will be launched twice in the pipeline:
First it will be launched as the first job and write the current time (= starting time of the pipeline) in the log file.
Finally it will be launched after all other jobs are completed and will append the log file with the end time of the pipeline.