remove top-level packaging scripts
The long term development goal of LAL is too split it up into
multiple packages that can be independantly released, when required. The
top-level packaging scripts are therefore against this effort. As there
are no plans to create monolithic lalsuite packages this patch removes
the packaging scripts that are adding an unnecessary maintenance burden.