specify the linker to use when condor compiling
the linker should be obtained from querying the compiler, using the
command "gcc -print-prog-name=ld", but currently condor returns a series
of error messages when this is done. hard coding the linker works round
this problem, and as condor will only work correctly when using the
system linker, which is always at /usr/bin/ld on supported systems, this
isn't as bad as it initially sounds.