set LIBTOOLIZE=true before running autoreconf
autoreconf from Ubuntu 9.10, and probably also from recent Debian
versions, insist on calling libtoolize with the --copy option which
can overwrite the libtool scripts. this can lead to unexpected build
failures, to work round this "feature" the LIBTOOLIZE variable is set to
point to the "true" executable which causes the running of libtoolize to
be skipped. this will not effect the vast majority of users and those it
does effect will know how, and when, to run libtoolize.