Changes to plotsnrchi: added plot with newsnr contours and made histograms logarithmic
The plot with 'contours' of slide density is removed. It is anticipated this will be replaced with a likelihood plot (or several) with a better
density estimation and smoothing.
Plot with newsnr contours is analogous to the effsnr contour plot. The contour values are 6,8,10,12.

The y-axes of the histograms were made logarithmic to make tails of distributions more visible.

Much general housekeeping, removal of obsolete 'variable denominator' statistic and renaming of 'constant denominator statistic' as effective snr.

First steps towards incorporation of 'autocorrelation chisq'.

Some changes to the .ini file and/or arguments passed to the script are necessary.

A sample output can be seen here
https://atlas3.atlas.aei.uni-hannover.de/~tdent/LSC/tests/H1-plotsnrchi_TOMS_ALLINJ_SUMMARY_PLOTS_CAT_4_VETO-933379143-604944.html
produced by the following command:

plotsnrchi --ifo-times H1 --gps-start-time 933379143 --gps-end-time 933984087 --snr-chisq --snr-chisq-lines --lines-effsnr --lines-newsnr
--denominator-constant 250. --chisq-index 6. --hist-effsnr --hist-newsnr --snr-rsq --rsq-coeff 0.0075 --rsq-max-snr 12 --rsq-pow 1.124764
--rsq-threshold 0.0002 --cache-file ../s6_lowmass_ihope.cache --found-pattern SIRE_INJECTIONS_*_FOUND*SECOND_*_*INJ_CAT_4_VETO --trig-pattern
SIRE_SECOND_*_PLAYGROUND_CAT_4_VETO --slide-pattern SIRE_SLIDE_SECOND_*_FULL_DATA_CAT_4_VETO --user-tag TOMS_ALLINJ_SUMMARY_PLOTS_CAT_4_VETO
--enable-output --output-path .

Unfortunately the plotting does not seem to be finely enough adjustable to produce newsnr contours that are continuous between the two parts of the
function to the last pixel. Mind the gap.