don't assume a flat namespace
some platforms, such as Mac OS X 10.4 and above, use a two level
namespace which results in _much_ faster symbol lookup times. the way
the lalDebugLevel variable is used assumes a flat namespace, therefore
this will not work correctly on platforms that don't use a flat
namespace, this patch removes the assumption that the namespace is flat.