comment out LALUserVarGetProcParamsTable
the LALUserVarGetProcParamsTable() function uses the
LIGOMetadataTables.h header, in the process of perform the split of the
inspiral code into a separate package all the metaio related code needs
to be moved into lalmetaio. this function is only used in lalapps_ninja
therefore it is being commented out until a better solution can be
found.