plotsnrchi: checks if any chisq_dof column contains zeros and refuses to plot corresp chisq
previously, it had been assumed that all chisq and chisq_dof columns would be populated with nonzero
entries. However, if some chisq (standard/bank/cont) is not calculated there is a 0./0 result when
finding the reduced chisq and the plot will either be nonsensical or fail. Added a check: if some class
of triggers (zerolag/slide/inj) exists and their chisq_dof is zero (or 1, for standard chisq) give a
warning. If the plotting arguments try to plot the chisq, return an error.

It had also been assumed that rsq was always calculated and cut on. But if plotsnrchi is run on files
with only bank/cont chisq, this is not the case. Added appropriate tests before reporting the rsq cut
output.