hard code python version in lalapps.spec
the autotools on CentOS 5 have an unpatched security hole that leads to
a potential "man in the middle" attack when preparing releases. as such
the next lalsuite releases will be prepared on another system which has
been patched for this issue. the spec file points to the python version
found during configuration, which may not be the same python version as
CentOS 5, therefore hard code this for the time being.