remove gpsepoch global variable
the variable was just a LIGOTimeGPS copy of the GPSStart command-line
option, so GPSStart is now used directly.

for simplicity, this patch also converts the GPSStart, GPSEnd, and
trigstarttime elements of the command line options structure from INT4 to
LIGOTimeGPS.  in principle this allows non-integer start and stop times to
be given on the command line, but for now this feature is disabled using
conditionals in the command-line code (non-integer values *should* work,
but this is not tested).

this patch also adds initializers to the .ht and .ht_V elements of the
global variable structure.  .ht didn't need to be initialized before, but
with the addition of support for Virgo it now must be to prevent a free of
the uninitialized pointer.