allow params_func to return None
In LikelihoodRatio and friends classes, a None return value from the
paramsfunc is passed out as None for the likelihood ratio, confidence,
whatever.  The goal is to allow the code to gracefully handle coincs for
which no values of the P(background) and P(Gw) probabilities can be
computed but for which this is not an error, for example coincs that have
had too many events vetoed to compute physical parameters.

For cases where the inability to compute these parameters is an error, the
params_func implementations must now raise exceptions themselves.