Script to assist moving LSD documentation to doxygen
Use barely incomprehensible regexes to match #if/#endif blocks
and comments, then checks to see if they contain <laldoc> directives.
If so, convert then to doxygen comment blocks, and try to fix up
embedded LaTeX.

Also, can run a check to see if the doxygenated file produces the
same output when run through the C preprocessor.