Multitute of TwoSpect improvements
Rather than slowly walk toward a stable solution, randomly
choose R values to find stable solutions. Added TwoSpect specific
functions to cdfwchisq since it always uses variables with 2 dof
and zero non-centrality parameter. Replaced goto's with a while
loop and if-else statements. Moved memory allocation for templates
out of a nested for-loop in TwoSpect.c. Try removing unnecessary
calls to numericFAR() in the signal parameter estimate part of the
code and only update to a better position if the FAP is smaller
(i.e. the R value is more significant)