Version bump
Also, I forgot to add in the previous commit that if a negative
root (negative R) is found, for the FAR threshold, then we try a
new starting position (randomly) and see if Newton's method
converges properly. This is limited to 10 tries. This is only
valid when the FAR is <<50%, say 10% or less.