Added functionality to TwoSpect
Can turn on/off calculation of R threshold value. In "off" state,
candidates' FAP are compared against the FAR as chosen by the
user. In the "on" state, the FAR threshold R value is calculated
for each template until a candidate is found with an R value
higher than the threshold. Then candidates continue to be ranked
by FAP.

Also added here a test to see if we can compute FAP for high
values of R using a least-squares-fit to a straight line.