dbtables.get_connection_filename():  work around "permission denied" errors
I'm getting random non-persistent "permission denied" errors on the CIT
cluster when working with database files on /usr1.  The errors are
reproducible in the sense that I always see them when I run certain
programs, but they are non-persisent in the sense that subsequent checks of
the relevant files and filesystems show no permission problems.

I don't know what causes this and I don't really care, so this patches
allows non-persisent "permission denied" failures to be worked around the
same way non-persisent "no space on device" failures are worked around.