glue.ligolw.dbtables:  rework signal trapping
This patch implements a signal handler that is turned on when scratch files
are in use and arranges to delete them and do a clean interpreter shutdown
when SIGTERM and SIGTSTP are received.  This should prevent scratch files
from being leaked on Condor clusters when jobs are killed or evicted.

The IOTrappedSignal exception is removed, and instead of raising that
exception the code now resends any intercepted signals to itself after file
I/O is completed.  This change is needed to interact properly with the file
clean-up signal handlers.