Added a jitter option to lalapps_binj.
This option allows for a random offset to be added to an injection time.
Additionally, if a min and max frequency option is specified to be the
same value, that exact value will be set at the center frequency.
(cherry picked from commit 75e29db583956177ca23ee9a98d8011a22cb7b9f)