pylal:  remove improper ligolw_add imports from programs
for some reason (probably improper cut-and-pasting) a whole pile of
programs that do not use ligolw_add at all are importing the module from
glue.  this is confusing my efforts to clean up sngl_inspiral ID remapping
because it makes my searches for code that depends on ligolw_add produce
false positives.  this patch removes the imports from said programs.