Added useful BoundedKDE class to pylal/kdes.py.
Can be used to compute kernel density estimates from distributions of points
where the domain should be periodic and/or bounded in some dimension(s) (for
example, on the sky, declination is bounded, and right ascension is peroidic).
Has a utility method that estimates the quantile of a sample under a greedy
algorithm that ranks the points by KDE-estimated density.