Fix fencepost error in ST4 in rare corner case
When fRef != 0, it can occasionally happen that an adaptive integration
stride drops below fmin, but the last evenly sampled interpolation step
is slightly above fmin. This can cause a fencepost error which
leads to nonsensical dynamics (and hence waveform) in the first sample.
This patch removes the fencepost error in such a case.