glue/ligolw/utils/table_utils: allow experiment times function to fail
The get experiment times function looks for the start/end times of the
segment summary produced by segments_from_cats. However this will be
absent if vetoes are not applied, eg in CAT1 coincidence jobs. Hence
allow this to pass with Value Error if the segment summary table is
absent and assign null values - nb CAT1 results are not used anyway