Fix incorrect condition for recomputing noise moments in sbank
This results in a tremendous speedup for placing TaylorF2RedSpinTemplate
using the metric, also making run time essentially independent of low
frequency cutoff.

The noise moments are recomputed at any given frequency resolution if
the high-frequency cutoff exceeds the highest frequency up to which the
noise moments had been previously computed.

The comparison between the length of the PSD and the length of the
old noise moment vectors was broken because it did not account for the
shift in indices between the PSD and the noise moments due to the low
frequency cutoff.

The condition was:

    len(PSD) > self.bank._moments[df][0].length

but it should be:

    len(PSD) - self.bank.flow / df > self.bank._moments[df][0].length