Added Quadrupole-Monopole and Tidal functionality to TaylorF2, ChooseXDWaveform. Added --eos command line option for lalinference. Added Yagi-Yunes I-Love-Q relation to calculate quadrupole-monopole parameters from love numbers.
NOTE: Edited LALInspiralWave.c because it calls ChooseTDWaveform and needed to pass a value for the neutron star eos. However, this eos is set to LAL_SIM_INSPIRAL_EOS_NONE and IS NOT read from the command line. I do not know what this code is for and therefore cannot edit it further.