pulsar_parameter_estimation_nested: Allow variable speed of gravitational waves
The pulsar parameter estimation code can now take in a variable speed of
gravitational waves. This should be input via the prior file and be a
number between 0 and 1 that is (1 - the fractional GW speed compared to the speed
of light) e.g. 0.001 means that the speed of GW is (1-0.001)c = 0.999c.