Redmine PR #953: Lalapps_inspiral flag --disable-filter-inj-only doesn't work if no injections are present
In the case that lalapps_inspiral is run with an injection file and --disable-filter-inj-only it will not analyse anything if no injections are present within the analysis window (+ safety zone). This is reasonably simple to fix, although I don't understand some of the logic around this point (why is a flag taking binary values rather than 1 or 0?).