snglcoinc.CoincParamsDistributions:  track counts and PDFs separately
add separate BinnedArray attributes containing the PDF data, and have the
.finish() method populate them and leave the raw counts alone.  also record
the PDFs in XML serializations so that .finish() doesn't necessarily need
to be called.

these features have been added in an ad hoc way in several applications.
for example the gstlal-based inspiral pipeline implements much of this in
the DistributionStats class, and the excess power and string cusp pipelines
provide the option of working with likelihood data in "counts" or "pdf"
form via command line options (leaving it up to the user to remember which
of the two they have in their .xml files).  implementing this here will
allow clean ups in those codes.  by preserving both the bin counts and the
PDFs, finally the .__iadd__() method knows how to weight the PDFs when
combining them.