This patch changes the PhenomP code in two ways:
As pointed out by Alejandro:

- It fixes a typo in the rotation matrix used in the function
  XLALSimIMRPhenomPCalculateModelParameters to compute the internal
parameter alpha0 from the spin components provided by the user. Note
that defining alpha0 amounts to performing a global rotation around the
total angular momentum of the system so it is a free parameter in the
intrinsic description of the PhenomP model (it redefines the origin of
the azimuthal angle phi in the frame where J is z). For the purpose of
interfacing the model with ChooseFDwaveform which requires 6 spin
components, these components are used to fix alpha0 but the previous
implementation resulted in a parametrization of alpha0 in terms of these
components which was very unnatural due to the typo.

- It replaces Taylor expanded expressions for the cos and sin of the
  inclination angle in the function WignerdCoefficients by non expanded
versions. This will barely affect the waveform inside the calibration
range of the model but the previous expanded expressions led to
divergencies in the waveform (pointed out by Salvatore Vitale) in some
regions of parameter space that are way outside of the calibration range
but that one could still be interested in probing (or not avoiding) for
instance with PE codes.

- Also updated the unit tests in test/PhenomPTest.c to agree with the
  current code.