Restore correct distance units after BAYESTAR MCMC
I scaled all of the distances in the problem to the maximum SNR=1
horizon distance amongst all of the detectors. This was harmless for the
Gaussian quadratures version of BAYESTAR because the distance was a
marginal parameter. However, the MCMC version has an option to save the
sample chains, so it is useful to save the output in Mpc.