Correct tC, provide t(f) and f(t) for SEOBNRv2ROMSingleSpin
Correct the waveform phase by a time shift such that the peak in the TD
coincides with the origin of time as defined by the epoch.

Provide new functions XLALSimIMRSEOBNRv2ROMSingleSpinTimeOfFrequency()
and XLALSimIMRSEOBNRv2ROMSingleSpinFrequencyOfTime() for computing the
time as function of frequency (relative to the waveform peak at t=0) or
the frequency as a function of time from the FD phasing.