rapidpe: Don't use LAL time series in rho_lm loop
The mode time series handed to the network log likelihood function was using LAL timeseries. This invoked a copy of the sliced numpy array (ts.data.data member) instead of a view since (we suppose) the 'array' object did not actually own the data and thus could not provide a copy. This is ameliorated by eschwing the LAL time series in favor of numpy arrays well before the loop, and moving the epoch of the time series (now ensure to be the same for all detector) to the time grid points. The construction of the time grid points is also moved outside the likelihood definition so as not to be redundnantly recreated.