Fr.c: Better differentiate failure modes
frgetvect will not differentiate between the inability to retrive a vect struct because it is not present or because the span requested is not commensurate with the span present in the frame. This patch adds additional error checking and raises a ValueError if the requested span is not compatible with the span of the frame.