made waveform string parsing routines more robust
The waveform string parsing routines used strstr to find the longest
matching substring, but were prone to match the wrong substring when
a typo in the waveform was made.  So, for example, the string
"SEOBNRv2_ROM_SinglSpin" had been considered valid because "SEOBNRv2"
matched the first part.  This adds more robust routines that demand
that the whole waveform string is valid.  Also, comparisons are done
in a case-insensitive sense.

The original routines remain (with deprecation warnings) but are
now made case-insensitive.  They should be deleted in an upcoming
release.