migrate btlwnb docs from lalapps to lalsimulation
the documentation for the excess power pipeline in lalapps contains the
writup of the algorithm by which band- and time-limited white-noise burst
waveforms are generated by XLALGenerateBandAndTimeLimitedWhiteNoiseBurst().
this patch moves that documentation from the excess power docs in lalapps
to the LALSimBurst.c documentation in lalsimulation, to live within
XLALGenerateBandAndTimeLimitedWhiteNoiseBurst().

part of the documentation for these waveforms (the big related to
normalization) belongs with XLALMeasureEoverRsquared(), so we place it
there.