Make 'astro mollweide' RA axis consistent across Matplotlib versions
A matplotlib < 1.2.0 code path had caused 'astro mollweide' axes to be
presented from 12h to -12h on old versions of Matplotlib, compared with
24h to 0h with modern versions of Matplotlib.

I just tested this on an SL6 machine (Matplotlib 1.0.1) and a Debian
Wheezy machine (Matplotlib 1.1.1rc2) without that code path, and the
axes were drawn correctly from 24h to 0h.

Maybe this special code path was necessary for Matplotlib < 1.0. Anyway,
it seems to work now, with the correct 24h to 0h axes regardless of
Matplotlib version.