BAYESTAR internal sign convention now conforms with FINDCHIRP
The unit test for the signal amplitude model is now much more elaborate,
anchoring the phase convention on injecting a signal with
XLALSimInspiralFD / XLALComputeDetAMResponse. This demonstrated that there
was a mismatch in the signal amplitude model between the code and the
BAYESTAR paper, which this patch corrects.

This patch also corrects a mismatch in the autocorrelation likelihood
between the code and the BAYESTAR paper.

The *default* phase convention for the high-level interface is still
anti-FINDCHIRP (conforming with gstlal and MBTA).

Fixes #2354.