Simplify hot loop in IMRPhenomD
Some calculations related to the reference frequency are done
inside the loop over frequency bins, but they do not depend on
the loop counter. This moves them out of the loop. It also
declares a few more variables as const. For (~10,~10) binaries
with random spins starting at 20 Hz, this change produces an
average ~30% speedup on my laptop.