Implement SkyMaxed Overlap for Precessing Templates
This patch moves around some of the class structure in
waveforms.py. The main highlight is that (i) The base class is now
called AlignedSpinTemplate, instead of Template, (ii) subclasses don't
have to implement the waveform calls. SBank will call into ChooseFD if
the approximant is already FD, or SimInspiralFD is the approximant is
TD, (iii) An attempt was made to unify functions like _get_dur() which
were ad hoc-ly defined for each approximant. I've left the definitions
of these functions unchanged here to help testing, but it probably
makes sense to eventually merge them. The code gives identical banks
for PhenomC and SEOBNRv2_ROM before and after the application of this
patch in a small BBH test case.