Added backwards integration.
XLALSimIMRSpinEOBWaveformAll takes an additional parameter fOut: if fOut<fmin, integrate backwards in time from t=0, with the stopping condition of Omega=fOut/2. Does not attach to the waveform with forward-integration, only produces the t<0 part. The two pieces, t<0 and t>0, are to be combined in post-processing.

Beware: preliminary commit, work in progress.
Even if the dynamics seems to be smoothly extended, there are still sign flips missing for the waveform.