sbank: don't truncate IMR waveforms too early
The current code uses the SEOBNRv2 ringdown frequency as the final freq
of all IMR waveforms. For some parameters, this can truncate the
waveform frequency series (used for computing matches) near or at the
ringdown. In regions of parameter space where waveforms are very short,
this can distort the waveform and produce visible holes in the bank.
This patch forces the final frequency of all IMR approximants to be the
maximum frequency used for match computation.