E@H: delay execution of app if not enough memory is available
The BOINC app can fail because of not enough memory at startup when a lot of tasks start at the same time or some non-BOINC program is using all the available memory and BOINC doesn't detect that fast enough.
Now our app calls boinc_temporary_exit() to signal the Client to delay the execution. This is a C++ interface so we also need a C wrapper to access it.
Our BOINC wrapper also still uses the old LAL error handling which has a special error value if an error occured in an XLAL function. This was not handled in the BOINC LAL errorhandler previously but is needed now to detect out-of-memory errors from XLAL functions.