sbank: implement variable low-frequency cutoff
This introduces the ability to optimize the low-frequency cutoff of each
template to be as large as possible (thus avoiding having very long template
waveforms) with the constraint of a fixed SNR loss relative to when the
template is computed using the minimum given low-frequency cutoff. The
per-template cutoff frequency is then recorded in a configurable column of the
single-inspiral table.