Change PSD extrapolation to INFINITE in LALInferenceReadData  - fixes #4911
From Carl Haster:
When reading in a PSD from a txt file this patch ensures that any PSD
values outside the frequency space defined in the file is set to
INFINITY.

I have checked that the PSD files that are produced are read in
correctly when used in lalapps_compute_roq_weights.py, i.e. the produced
Numpy array with the PSD is filled with numpy.inf values where
appropriate.

This patch also adds a check in the PSD plotting function in
CBCBayesPostProc to make sure that INFINITE PSD values are ignored.

Finally, the plotted PSD now starts at the f_low used in the analysis,
as opposed to the default 30 Hz that was hardcoded previously.

This isn't mission critical for our O2 analysis, but would probably be
nice to have pretty soon anyways