Do not import exceptions module
Required for compatibility with Python >= 2.7 and 3.

According to the Python 2.7 documentation:

"The exceptions are defined in the module exceptions. This module never needs to be imported explicitly: the exceptions are provided in the built-in namespace as well as the exceptions module."

The Python 3 manual makes no mention of the exceptions module.