bayestar: add uniform in comoving volume prior
It's disabled by default so that we do not accidentally change the
volume prior in the middle of an observing run.

The WMAP9 cosmological parameters are hard-coded. The prior is
implemented as a cubic spline approximation to the correction
factor relative to a DL^2 prior. This is simple, fast, accurate, and
allows us to use BAYESTAR's already highly tuned distance integrator.