bayestar_plot_volume: fix blank plots with certain Numpy versions
A coordinate singularity at the origin would cause the central pixel
to have a very large value, washing out the color map.

In the process of improving the unit tests, I also corrected the
normalization of the output of the volume renderer. It wasn't essential
for the normalization to be correct because the plot is auto-scaled
anyway. However, it's satisfying that it is correct now anyway.

Fixes https://bugs.ligo.org/redmine/issues/5591.